Streamlining process
This is the current work in progress of creating the custom bsp for the board. Making an easier to follow guide of how to make the custom device tree. Currently not bootable. Says it has to do with not being able to load the bootenv.txt and the device tree blob.